pyproject.toml
[tool.poetry]
name = "sberright"
version = "0.1.0"
description = ""
authors = ["Malvere <gatling1404@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.39.0"
aiogram = "^3.1.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
